Data-AI-Course/requirements.txt
# AI-Powered Workflows Bootcamp - Requirements
# Week 00 - Setup & Tools

# Core LLM SDKs
openai==1.12.0
anthropic==0.18.1

# LLM Frameworks
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.20
llama-index==0.10.11

# Machine Learning
transformers==4.38.0
torch==2.2.0
sentence-transformers==2.3.1

# Data & Utilities
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
numpy==1.26.4
pydantic==2.6.1

# Vector Databases
faiss-cpu==1.7.4
chromadb==0.4.22

# Development Tools
jupyter==1.0.0
ipykernel==6.29.2
notebook==7.1.0

# Testing & Quality
pytest==8.0.0
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# API & Web
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Utilities
tqdm==4.66.2
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
